COMPREHENSIVE BIGINT FIX: Convert ALL user_id/admin_id columns from INTEGER to BIGINT across all files
- database_testforwarder.py: Fixed user_id INTEGER  BIGINT (4 occurrences)
- referral_system.py: Fixed user_id/referrer_user_id/referred_user_id INTEGER  BIGINT
- vip_system.py: Fixed user_id INTEGER  BIGINT
- bump_service_testforwarder.py: Fixed user_id INTEGER  BIGINT (2 occurrences)
- product_price_editor.py: Fixed changed_by_admin_id INTEGER  BIGINT
- interactive_welcome_editor.py: Fixed admin_user_id INTEGER  BIGINT

This addresses the root cause: psycopg2.errors.NumericValueOutOfRange for Telegram user ID 6622951691

+++ bump_service_testforwarder.py
@@ -383,7 +383,7 @@ class BumpService:
             cursor.execute('''
                 CREATE TABLE IF NOT EXISTS ad_campaigns (
                     id INTEGER PRIMARY KEY AUTOINCREMENT,
-                    user_id INTEGER,
+                    user_id BIGINT,
                     account_id INTEGER,
                     campaign_name TEXT,
                     ad_content TEXT,
@@ -464,7 +464,7 @@ class BumpService:
                 CREATE TABLE IF NOT EXISTS ad_performance (
                     id INTEGER PRIMARY KEY AUTOINCREMENT,
                     campaign_id INTEGER,
-                    user_id INTEGER,
+                    user_id BIGINT,
                     target_chat TEXT,
                     message_id INTEGER,
                     sent_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
